FW1_04_17

+++ project/korra_common/packages/wdcore_avatar/sbin/CalTransfer.sh
@@ -12,8 +12,8 @@ echo $timestamp ": CalTransfer.sh Start" $total $size $state $exist >> /tmp/back
 while [ `cat /tmp/transfer_state | sed -n 's/.*=//p'` == "running" ]; do
 	exist=`cat /tmp/transferExistSize`
 
-	total=`cat /tmp/transfer_size_total`
-	size=`cat /tmp/transfer_size`
+	total=`cat /tmp/transfer_size_total | awk -F= '{print $NF}'`
+	size=`cat /tmp/transfer_size | awk -F= '{print $NF}'`
 	state=`cat /tmp/transfer_state`
 
 	#timestamp=$(date "+%Y.%m.%d-%H.%M.%S")
@@ -32,8 +32,10 @@ while [ `cat /tmp/transfer_state | sed -n 's/.*=//p'` == "running" ]; do
 			transferDoneSize=`expr ${transferDoneSize} + ${divide}`
 			#echo $timestamp ": CalTransfer.sh new transferDoneSize:" ${transferDoneSize} >> /tmp/backup.log
 			#echo $timestamp ": CalTransfer.sh new divide:" ${divide} >> /tmp/backup.log
-			if [ "${exist}" -gt "${transferDoneSize}" ]; then
+			if [ "${exist}" -gt "${transferDoneSize}" ] && [ "$transferDoneSize" -le "$total" ]; then
 				echo "transferred_size_in_bytes=""$transferDoneSize" > /tmp/transfer_size
+			else
+				echo "transferred_size_in_bytes=""$exist" > /tmp/transfer_size
 			fi
 			if [ "${transferDoneSize}" != "" ]; then
 				echo "${transferDoneSize}" > /tmp/transferDoneSize
@@ -44,20 +46,18 @@ while [ `cat /tmp/transfer_state | sed -n 's/.*=//p'` == "running" ]; do
 		transferDoneSize=`cat /tmp/transferDoneSize`
 		if [ "${transferDoneSize}" == "" ]; then
 				transferDoneSize=0
+		elif [ "${transferDoneSize}" -gt "${total}" ]; then
+				transferDoneSize=${total}
 		fi
 		
 		if [ -f "/tmp/runningBackupDst" ]; then
 			backupFolder=`cat /tmp/runningBackupDst`
 		fi
 		if [ "${backupFolder}" != "" ]; then
-			#rsyncCt=`ps aux | grep rsync | grep rv | wc -l`
-			#if [ "$rsyncCt" -le 1 ]; then
-			DoneSize=`rsync -rv "${backupFolder}" | grep "total size is" | awk '{print $4}' | sed -n 's/,//gp'`
-			#RsyncExecmd="du -s \"${backupFolder}\" > /tmp/DoneSize"
-			#eval "$RsyncExecmd"
-			#folderSizeKB=`cat /tmp/DoneSize | awk '{print $1}'`
-			#DoneSize=`expr "${folderSizeKB}" \* 1024`
-	
+			DoneSize=`/bin/du -s -b "${backupFolder}" | awk '{print $1}'`
+			#DoneSize=`expr "${CapacityCal}" \* 1024`
+			#DoneSize=`rsync -rvn "${backupFolder}" | grep "total size is" | awk '{print $4}' | sed -n 's/,//gp'`
+			#echo $timestamp ": CalTransfer.sh DoneSize" $DoneSize >> /tmp/backup.log
 			if [ "$DoneSize" == "" ]; then
 				DoneSize=0
 			fi
@@ -70,7 +70,10 @@ while [ `cat /tmp/transfer_state | sed -n 's/.*=//p'` == "running" ]; do
 			else
 				transferredSize=`expr $DoneSize - $OrigSize + $transferDoneSize`
 			fi
-			
+			#echo $timestamp ": CalTransfer.sh OrigSize" $OrigSize "transferDoneSize" $transferDoneSize >> /tmp/backup.log
+
+			#echo $timestamp ": CalTransfer.sh transferredSize" $transferredSize >> /tmp/backup.log
+
 			if [ "$transferredSize" -gt "$total" ]; then
 				transferredSize=$total
 			fi
@@ -84,15 +87,15 @@ while [ `cat /tmp/transfer_state | sed -n 's/.*=//p'` == "running" ]; do
 	
 	sdpct=0
     sdpct=`expr \`cat /tmp/transfer_size | cut -d "=" -f 2\` \* 100 / \`cat /tmp/transfer_size_total | cut -d "=" -f 2\``
-    echo "$sdpct" > /tmp/sdpct                            
+    echo "$sdpct" > /tmp/sdpct
     if [ $sdpct -ge 25 ] && [ $sdpct -lt 50 ]; then
 		echo "42;25" > /tmp/MCU_Cmd                                                                
     elif [ $sdpct -ge 50 ] && [ $sdpct -lt 75 ]; then
 		echo "42;50" > /tmp/MCU_Cmd
-    elif [ $sdpct -ge 75 ] && [ $sdpct -lt 100 ]; then
+    elif [ $sdpct -ge 75 ] && [ $sdpct -le 100 ]; then
         echo "42;75" > /tmp/MCU_Cmd              
-    elif [ $sdpct -ge 100 ] ; then               
-        echo "42;100" > /tmp/MCU_Cmd
+    #elif [ $sdpct -ge 100 ] ; then               
+    #    echo "42;100" > /tmp/MCU_Cmd
     fi
     sleep 5
 done
@@ -100,7 +103,7 @@ done
 total=`cat /tmp/transfer_size_total`
 size=`cat /tmp/transfer_size`
 state=`cat /tmp/transfer_state`
-#echo "42;100" > /tmp/MCU_Cmd
+echo "42;100" > /tmp/MCU_Cmd
 timestamp=$(date "+%Y.%m.%d-%H.%M.%S")
 echo $timestamp ": CalTransfer.sh END" $backupFolder $total $size $state >> /tmp/backup.log
 
@@ -108,7 +111,4 @@ if [ -f "/tmp/getStorageTransferStatus" ] ;then
 	rm /tmp/getStorageTransferStatus
 fi
 
-#rsync_pid=`pidof rsync`
-#if [ "${rsync_pid}" != "" ]; then
-#	kill -15 ${rsync_pid}
-#fi
+
